transbot/requirements.txt
discord.py>=2.0.0
Flask>=2.0.0
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
gunicorn>=20.1.0
